+++ R/extract-png-info.R
@@ -18,8 +18,8 @@ color_types[[6 + 1]] <- "RGB + Alpha"
 #'
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-png_info <- function(raw_vec) {
-  res <- .Call('png_info_', raw_vec)
+extract_png_info <- function(raw_vec) {
+  res <- .Call('extract_png_info_', raw_vec)
 
   res$color_desc <- color_types[[res$color_type + 1L]]