+++ src/init.c
@@ -4,12 +4,14 @@
 #include <Rinternals.h>
 
 extern SEXP read_png_raw_(SEXP raw_vec_, SEXP fmt_, SEXP flags_);
-extern SEXP extrac_png_info_(SEXP raw_vec_);
+extern SEXP read_png_nara_(SEXP raw_vec_, SEXP flags_);
+extern SEXP extract_png_info_(SEXP raw_vec_);
 
 static const R_CallMethodDef CEntries[] = {
-
-  {"read_png_raw_", (DL_FUNC) &read_png_raw_, 3},
-  {"extrac_png_info_", (DL_FUNC) &extrac_png_info_, 1},
+  
+  {"read_png_raw_"   , (DL_FUNC) &read_png_raw_   , 3},
+  {"read_png_nara_"    , (DL_FUNC) &read_png_nara_    , 2},
+  {"extract_png_info_", (DL_FUNC) &extract_png_info_, 1},
   {NULL , NULL, 0}
 };